220320조별과제완료

+++ UglyShop2/WebContent/review/reviewDetailUser.jsp
@@ -30,53 +30,44 @@
     margin: 20px 0 0 20px;
   }
 </style>
-<!-- <div class="row mt-5"> -->
-<div class="col-md-8 mx-auto">
-  <div class="font-weight-bold mt-3 shadow p-3 mb-4 bg-light rounded">리뷰 상세보기 페이지</div>
-  <div class="review rounded">
-    <h4><c:out value="${review.reviewTitle}" /></h4>
 
-    <div class="info">
-      <p>글번호: <c:out value="${review.reviewID}" /></p>
-      <p>상품번호: <c:out value="${review.prodID}" /></p>
-      <p>작성자: <c:out value="${review.userID}" /></p>
-    </div>
-    <p class="date">작성일: <c:out value="${review.reviewDate}" /></p>
+
+<div class="col-md-8 mx-auto mt-3">
+  <div class="review">
+    <h4>제목: <c:out value="${review.reviewTitle}" /></h4>
+    <p>리뷰글번호: <c:out value="${review.reviewID}" /></p>
+    <p>작성자: <c:out value="${review.userID}" /></p>
+    <p>작성일자: <c:out value="${review.reviewDate}" /></p>
+    <p>상품번호: <c:out value="${review.prodID}" /></p>
+    <p>판매자: <c:out value="${review.farmID}" /></p>
     <p class="content"><c:out value="${review.reviewContent}" /></p>
     <c:if test="${userID eq review.userID }">
       <!-- board 참고해서 button위치 옮겨놓기 -->
       <c:if test="${review.reviewID ne null }">
-        <div class="edit" style="text-align: center; align-content: center">
-          <button type="button" class="btn btn-success" onclick="location.href = '<%=request.getContextPath()%>/reviewController2?cmd=edit&reviewID=<c:out value='${review.reviewID}'/>' ">수정</button>
-          <!-- button대신 a태그로 onlick(팝업창)과 cmd이동 둘다 가능  -->
-          <a
-            class="btn btn-danger"
-            type="button"
-            onclick="if((!confirm('정말로 삭제할까요?'))) return false"
-            href="<%= request.getContextPath() %>/reviewController2?cmd=delete&reviewID=<c:out value='${review.reviewID}'/>"
-            >삭제</a
-          >
-          <!-- a태그로도 가능 -->
-        </div>
+         <div class="edit" style="text-align:center; align-content: center; ">
+				<button type="button" class="btn btn-success"  onclick = "location.href = '<%=request.getContextPath()%>/reviewController2?cmd=edit&reviewID=<c:out value='${review.reviewID}'/>' ">수정</button>
+				<!-- button대신 a태그로 onlick(팝업창)과 cmd이동 둘다 가능  -->
+				<a class="btn btn-danger" type="button" onclick="if((!confirm('정말로 삭제할까요?'))) return false" href="<%= request.getContextPath() %>/reviewController2?cmd=delete&reviewID=<c:out value='${review.reviewID}'/>">삭제</a> <!-- a태그로도 가능 -->
+			</div>
       </c:if>
-    </c:if>
+     </c:if>
   </div>
+
   <!-- reply.farmID에 값이 있을때만 출력(ne가 !=와 같음) -->
   <c:if test="${reply.farmID ne null }">
-    <div class="reply rounded">
-      <div class="info">
-        <p>덧글번호: <c:out value="${reply.replyID}" /></p>
-        <p>상품번호: <c:out value="${reply.prodID}" /></p>
-        <p>판매자ID: <c:out value="${reply.farmID}" /></p>
-      </div>
-      <p class="content"><c:out value="${reply.replyContent}" /></p>
+  <form id="replyWrite" name="replyEdit" action="<%= request.getContextPath() %>/replyController" method="post">
+    <div class="reply">
+      <p>덧글번호: <c:out value="${reply.replyID}" /></p>
+      <p>판매자ID: <c:out value="${reply.farmID}" /></p>
+      <p>상품번호: <c:out value="${reply.prodID}" /></p>
+      <p><c:out value="${reply.replyContent}" /></p>
+      <p>리뷰글번호: <c:out value="${reply.reviewID}" /></p>
     </div>
+    </form>
   </c:if>
-  <div class="back">
-    <button onclick="location.href = '<%=request.getContextPath()%>/reviewController2?cmd=list' " type="button" class="viewList btn btn-primary">전체리뷰 보기</button>
-    <!-- 뒤로가기.리스트로 이동  -->
-  </div>
+
 </div>
+
 
 <jsp:include page="../includes/footer.jsp" />
 <script>
